backend/backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn>=0.28.0,<0.29.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.1,<3.0.0
sqlalchemy>=2.0.27,<3.0.0
alembic>=1.13.1,<2.0.0
asyncpg>=0.29.0,<0.30.0
psycopg2-binary>=2.9.9,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.9,<0.1.0
email-validator>=2.1.0,<3.0.0
fastapi-pagination>=0.12.17,<0.13.0
fastapi-cache2>=0.2.1,<0.3.0
fastapi-limiter>=0.1.6,<0.2.0
fastapi-mail>=1.4.1,<1.5.0
openai>=1.18.0,<2.0.0
anthropic>=0.18.0,<0.19.0
tenacity>=8.2.3,<9.0.0
python-dotenv>=1.0.1,<2.0.0
